requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pandas>=2.0.0
geopy>=2.3.0
haversine>=2.8.0
pytest-asyncio>=0.26.0